refactor: use API endpoints for assistant creation/edit
