+++ src/routes/settings/(nav)/assistants/[assistantId]/edit/+page@settings.svelte
@@ -1,11 +1,10 @@
 <script lang="ts">
-	import type { PageData, ActionData } from "./$types";
+	import type { PageData } from "./$types";
 	import { page } from "$app/state";
 	import AssistantSettings from "$lib/components/AssistantSettings.svelte";
 
 	interface Props {
 		data: PageData;
-		form: ActionData;
 	}
 
 	let { data, form = $bindable() }: Props = $props();
@@ -13,4 +12,4 @@
 	let assistant = data.assistants.find((el) => el._id.toString() === page.params.assistantId);
 </script>
 
-<AssistantSettings bind:form {assistant} models={data.models} />
+<AssistantSettings {assistant} models={data.models} />
